refactor(adapter-tests): invert conformance dependency — cases export, adapters import (#1194 review)
The previous shape had `adapter-tests/__tests__/template-primitives-conformance.test.ts`
import HonoAdapter and GoTemplateAdapter directly, then loop over both.
That couples the test infrastructure to every concrete adapter — adding
a new adapter requires editing `adapter-tests`, which constrains
adapter authors' freedom and inverts the dependency direction.

Mirror the existing `runJSXConformanceTests` pattern instead. Each
adapter package owns its own conformance test file and imports the
shared cases + harness:

  packages/adapter-hono/__tests__/template-primitives-conformance.test.ts
  packages/adapter-go-template/src/__tests__/template-primitives-conformance.test.ts

`@barefootjs/adapter-tests` exports:
  - `runConformanceSuite` (the harness, refactored to take a single
    adapter rather than an array)
  - `templatePrimitiveCases` + `TemplatePrimitiveCaseId` +
    `runTemplatePrimitiveCase` (the case data + standard runner)

Adding a new adapter is now: create one test file in that adapter's
package, declare its skip set, done. No central registry to update.
Adding a new case is: edit `packages/adapter-tests/src/cases/template-primitives.ts`,
every adapter picks it up the next test run.

The deleted file `packages/adapter-tests/src/__tests__/template-primitives-conformance.test.ts`
is replaced by the two per-adapter files. Test counts shift accordingly:

  adapter-tests:    -8 (the central file)
  adapter-hono:     +4 (Hono cases all pass)
  adapter-go-template: +4 (all skipped until #1188)

Pre-existing csr-conformance.test.ts / ssr-hydration-contract.test.ts in
`adapter-tests` still import HonoAdapter as a reference adapter — that
is a different pattern ("compare other adapters' HTML output to Hono
reference"), out of scope for this change.
